Add files via upload
line charts are excellent for showing trends and changes over time, making them suitable for time series analysis of air quality data. Heatmaps, on the other hand, are better suited for visualizing spatial and multivariate patterns, helping to identify spatial trends and correlations in air quality measurements. The choice of visualization technique depends on the specific insights you want to gain from your air quality data and the nature of the data itself